+++ lib/mangoapps/modules/learn.rb
@@ -2,6 +2,7 @@
 
 require_relative "learn/course_catalog"
 require_relative "learn/course_categories"
+require_relative "learn/course_details"
 require_relative "learn/my_learning"
 
 module MangoApps
@@ -10,6 +11,7 @@ module MangoApps
       # Include all learn sub-modules
       include MangoApps::Client::Learn::CourseCatalog
       include MangoApps::Client::Learn::CourseCategories
+      include MangoApps::Client::Learn::CourseDetails
       include MangoApps::Client::Learn::MyLearning
     end
   end